requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-parler==2.3
django-rosetta==0.10.0
gunicorn==23.0.0
idna==3.8
packaging==24.1
pillow==10.4.0
polib==1.2.0
psycopg2==2.9.9
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
